Add dynamic menu for AIS targets to show/hide CPA line
I used/tested this function already for 2 weeks while sailing. Found it
very usefull to estimate where and how CPA will happen with target (ahead/aft/starboard/port)
It uses the original CPA line which is used in case of an alert, and is updated frequently.
Menu is only shown when we have a valid CPA, and CPA line is only shown as long as CPA is valid.
It works target specific, so it is possible to show the CPA-line for several targets in parallel