Right click on own vessel when a route is active and there is a cross track error to provide a 'Zero XTE' option.
This option will use the same logic as getting to the first waypoint of a route by creating a virtual waypoint to get a route from the current postion to the active waypoint.

This is for FS#1563

This change allows the autopilot to follow a direct route to the next waypoint without haveing to converge on the original route first. It is very useful when sailing or motor/sailing.