Correct hardware accelerated panning of cm93 charts in non-quilted mode
Commit 40a80feb32daa7633b2f234367bef2c062c64d2d is a good optimization because
the chart cell was selected again in setvpparms making the removed logic redundant,
but broke the needed invalidation.

Instead move the logic from SetVPParms into AdjustVP so the viewport can be
invalidated if needed.  This is even faster than before because AdjustVP is called
less than SetVPParms.

Since the return value of AdjustVP is never used, some logic was removed.