optimize route distance calculation when adding new points
This greatly speeds up the loading time after a crash with a large number of route points
in a track to add.  For example if my computer ran out of power after a few days,
the next boot of opencpn trying to apply navobj.xml.changes would take about a minute
which is a very long time.  Now it is less than a second.

Similar logic could be applied for deleting points or other operations, but
maybe it's not really useful.

Conflicts:
	src/NavObjectCollection.cpp