Partially fix FS#1465 by removing scan limit.
There was an implicit threshold in `ChartDatabase::DetectDirChange()`,
empirically set to 10'000 files, that triggered speedy termination
with a result that means "This directory contents has changed, it needs
to be re-processed again". Therefore, instead of improving performance
by avoiding some large and unnecessary job, it resulted in doing
even larger, much more superfluous and complex processing (which
also needs to be fixed, by the way).