correct cog and sog filters
before the filters were run on any nmea message, even ones without
cog or sog updates.  This resulted in poor filtering and incorrect
initial filter values until the buffer filled.

The course up filter was running off the already filtered cog which
is not expected by users, this is corrected.

The status bar which is expensive to update is now not updated
unless the relevant data was in the message.