support compressed charts
This adds support for xz compressed charts if liblzma is available
Tested with kap, s57, and cm93 charts.

It does not perform any compression, for example compressing senc
s57 parsed charts but this may prove to be very useful.

This saves a lot of disk space with negligable performance loss.
It might be faster to load charts if the disk is slow (sd card etc)

Currently it completely decompresses s57 charts just to read the
header which is not ideal, but typically performed only once
because the senc is then generated.

My charts are all at least half the initial size, and many s57 charts
are around 25% of the normal size.  cm93 charts compressed to 60%