Fix crash when exporting route to track
Code used a tracklist pointer for a route list instead of a track list.  This caused
a crash when converting a track (generated by weather routing) into a route using
the track and route dialog.

The fix for this is in OnTrkRouteFromTrackClick.  This patch also fixes a similar
issue in OnTrkMenuSelected but I do not know how to invoke this code, so please
verify that this is correct before merging.