Make tide rectangles' displayed size depend screen resolution
The size of tide rectangles (the rectangular "icons" drawn on the
chart to show tide height at tide stations) was previously hard-coded
in terms of pixels, and so the rectangles become too small to be
readable on high-resolution displays. The changes in this commit make
the size of the tide rectangles depend on screen resolution, to be
readable on any display.

The changes also introduce a parameter TideRectangleScale, default
value 100, which can be set in the opencpn configuration file to allow
users to modify the scaling of the tide reactangles, by analogy with the
existing CurrentArrowScale parameter for adjusting displayed size of
current arrows.

Cf. http://www.cruisersforum.com/forums/f134/opencpn-scaling-problems-with-high-resolution-displays-179462.html