Always show zero line on current graphs
In some circumstances, current graphs would be displayed without an
explicit "0" line, which would make it difficult to see the precise
times of slack current. This commit arranges the graph's  vertical
coordinates so that the "0" line of slack current is always displayed.

This implements (my own) flyspray feature request FS#2119.