Scale tide and current graph windows depending on screen resolution
Even in Beta 4.5.213, tide and current graph windows (instances of TCWin)
do not render well on high-resolution displays because the window size
is too small. This commit addresses this problem by scaling each TCWin by
a factor that depends on the screen resolution.

This commit also introduces a configuration file parameter
TideCurrentWindowScale, default value 100, which allows the user to
modify the scaling, along the lines of the existing parameters
CurrentArrowScale and TideRectangleScale.

Cf. http://www.cruisersforum.com/forums/f134/opencpn-scaling-problems-with-high-resolution-displays-179462.html#post2316692