This adds US ACE "Inland ENC Buoy Overlay" and "Inland ENC SW Pass Overlay" catalogs to chart_sources.xml in chartdldr_pi. The <name>, <location> and <dir> for each catalog is added to the chart_sources.xml in the OpenCPN package. The catalogs are maintained by the US Army Corps of Engineers. There is one IENC overlay in each catalog. The Buoy Overlay adds the navigation buoys that are maintained and repositioned by the US Coast Guard. The buoys are relocated for changing river stages, shoaling, etc. These buoys are not shown the many river charts produced by the Corps, The Buoy Overlay is needed to display them in OpenCPN. The SW Pass Overlay contains newer sounding data for the Southwest Pass of the Mississippi River than is on the NOAA ENC. This overlay is of interest mostly for operators of deep draft vessels.
The chartcatalog.cpp and chartcatalog.h are revised to parse and display
the new catalogs. Also, a minor bug in the generation of records (lines)
within chartdldr_pi.dat is fixed for IENC and ENC charts. It was minor in
that the chartdldr_pi corrected the records for chart names that begin with a
a letter, but failed write correct records for chart names beginning with a
digit. So, it failed for the 3UABUOYS and 3UASW000 overlay chart names.
This was not a bug for RNC file names.