Plugin API 1.14 - Externalize AIS drawing
The BR24radar_pi group is adding (M)ARPA support for which we are sending ARPA targets to OpenCPN.

We would like to be able to draw these on the radar-only PPI windows, with the added bonus that we also see the AIS targets on the PPI display. This is also what the Navico plotters do.

This commit refactors ais.cpp in as minimal a fashion as possible so that it allows sharing of the AIS drawing code with the plugin(s).