Add updated licensing info (#757)
The data/copyright is, although updated, flawed since all the used
licenses (as of LICENSING) should go here. On the other hand, to
maintain this file is a downstream task which might be left out
at upstream.

The licensing seems mostly OK. The exception is the Apache-2.0
license which isn't GPL-compatible. Practically this is almost
only a android issue, those binaries could not be distributed
legally. Besides this, a single js-file in docs would be coped with
in downstream packaging.