Linux: plugin loading: Support data dirs in different locations.
The second and final part of making it possible to load plugins
installed in e. g. /usr/local when the main opencpn installation
lives in /usr.

This is not transparent: it works if and only if the plugin uses
the new GetPluginDataDir() instead of GetPluginDir() to get the
location of the plugin data directory. However, I see no way
around this. Having the possibilty is the best we can do.

By default, the patch locates data directories in ~/.local/share,
/usr/local and /usr. If the installation prefix isn't any of these,
it's also searched.