build: Preserve existing linker flags (#1144).
Current code resets CMAKE_SHARED_LINKER_FLAGS in several places,
effectively discarding the traditional build environment as of
LDFLAGS and also user-defined -DCMAKE_SHARED_LINKER_FLAGS=... on
the command line. This patch fixes this for linux builds, but does
not address windows or MacOS which suffers from the same problem.