Licensing: Cleanup #1100.
There are still leftovers of the gplv2 (as opposed to gplv2+) license.
This commit adds  a proper COPYING.gplv2 to the root directory for
symmetry, and also updates the runtime license file data/license.txt
to an unmodified copy of COPYING.gplv2 fresh from Free Software
Foundation, nuking the plain gplv2 statement in the original
data/license.txt.