cmake: Link to libGL.so rather than libOpenGL.so (#1157).
Explanation from https://github.com/NVIDIA/libglvnd/issues/94

Applications should link against libGL.so or libGLX.so.

libGL.so is a wrapper around libGLX.so, but from an application's'
perspective, it's the same as it would be for a non-libglvnd system.
Use libGL.so if you want to maintain binary compatibility regardless
 of whether a system uses libglvnd or not.

On a libglvnd-based system, you can link against libGLX.so instead. It
exports only GLX entrypoints, so to get OpenGL functions, you can
optionally link against one of the entrypoint libraries, typically
libOpenGL.so. It's optional because you could also just call
eglGetProcAddress for everything.

Note that the GL entrypoints all go through the same dispatch table, so
the same function is interchangeable no matter where you look it up --
libOpenGL.so, libGL.so, libGLES*.so, glXGetProcAddress, and eglGetProcAddress.