logging: Enable logging in all builds.
Make the log/don't log a runtime decision. To fully work, this
requires that the debug support is enabled also in release
builds. The idea is that this does not impose any serious performance
hit as long as debug logging isn't enabled in runtime using --loglevel.
This is definitely true for the new LOG_XXX macros.

Some wxWidgets libs are compiled without debug support. Use a
cmake try_compile() to identify them and disable debug logging.