AIS_Decoder, pluginmanager: Fix string overruns
The purpose of strncpy is to avoid overruns. AIS_Decoder uses the
strlen(src)  to compute the maximum size which doesn't make any
sense. pluginmanager uses the size of a pointer as maximum size
which is plain wrong.

Declare size of char arrays involved, and make a simple memcpy()
instead. This is safe and will always work as long as there is
enough data in src. If not, nothing will help.