glTexCache, glTextureManager: Fix localtime_r warning.
.wxWidgets emits a warning about using an unsafe localtime_r() wrapper.
The reason for this is that the symbol HAVE_LOCALTIME_R is undefined
when including the wx time.h file.

Now, HAVE_LOCALTIME_R is present in wx/setup.h so this should not happen.
Including wx.h before dylib.h fixes the problem, but I actually cannot
understand why.