Sound: Fix broken bells (#1277).
Despite the fix committed, the real cause for this bug seems to be
an oversight: while the sound code seems to work fine, it had to be
used the right way. The bells code was basically assuming the old sound
semantics, which caused the bug.

This also means to revert the earlier fix, which does not address the
bad usage pattern. Also, it adds some new problems which should be
avoided such as assuming that a sound factory is either synchronous
or asynchronous -- this is simply not true, the sound object is
configured after being created.