build: Get rid of hardcoded USE_WEBVIEW.
Current code is hard to maintain... Make a simple test if we can
use the webview library or not i CMakeLists.txt instead. Note:
this strategy does not scale, if we need to configure for more the
existense of one library we lose. But for now, this should be fine.