Build: Remove broken RPM generation (#1290).
The RPM package generation is broken, and the root is the underwhelming
CPACK capabilities in this respect. Proper RPM generation could be done
in a CI job based on a proper spec file. This would generate much better
packages and a proper srpm with lesser job.

With this patch, RPM-based platform defaults to generate simple
tarballs.