Restore 1f052c0 cmake wxWidgets detection (#1355).
This is actually a can of worms opened. Part of the problem is the
use of wxWidgets_USE_LIBS instead of wxWidgets_FIND_COMPONENTS which
does not work on modern cmakes, fixed in 1f052c0 but lost when
rebasing a later PR.

Another thing is the C code using #ifdef OCPN_USE_WEBVIEW. This flag
is not always correct, webview could be disabled due to e. g.,
missing backends even if configured. In this case wxUSE_WEBVIEW is
set to 0, otherwise 1. So the correct test is #if wxUSE_WEBVIEW,
augmented with a configuration HAVE_WEBVIEW flag.

Current code runs FIND_PACKAGE(wXwidgets) twice leading to all sorts
of problems. This is handled in #1298, but needs a point fix here.

Here is also problems with missing libraries when using GTK3,
handled by nasty fix.