build: Make the wxcurl lib visible (#1339).
The oesenc_pi plugin (at least) uses the wxcurl code. The normal way
for plugins is to duplicate libraries in their builds. However,
there cannot be two wxcurl libraries in the same program -- it
causes runtime errors whne trying to register events that
already exists.

Make the code visible for clients. This is an exception to the general
policy that plugins should bundle their deps dictated by how tightly
the wxCurl lib is coupled to the event loop.