ci: Add cloudsmith deployment, circleci builds + fixes.
Ensure that all artifacts have unique names so they can be
deployed in a common namespace. Make the deployment script
generic so it can be used on all builders.

Add configuration for circleci, similar to travis and re-using
most of the build scripts.

Also: clean up various code duplication in the ci/ directory.

The cloudsmith deployment relies on a CLOUDSMITH_API_KEY
variable which must be set in the build environment.
Without this the deployment is silently ignored.