UI: Display plugin status (#1519)
Plugins could be in one of four states:

  - Managed by the installer
  - System plugins, cannot be managed at all
  - Packaged plugins cannot be managed by installer but
    probably with the same tools as used to install them.
  - Ghosts are managed by installer, but there is an
    unused plugin with the same name, probably packaged.

The plugin metadata already contains if plugin is readonly
(i. e., packaged) or not. Add functions to determine if it
is a system plugin or a plugin which is shadowing a
packaged plugin.

Add an icon and some text to PluginPanel which shows status.

Closes: #1519