options: Add compatOS/compatOsVersion options.
We need a better alternative to the current pasrsing of /etc/os-version
to determine compatible platforms like Mint which should be treated as
Ubuntu. The main problem is that the OS version cannot be found this
way.

Add options to define the OS and OS version to use when determining
what plugins to load. There is no gui at this point, users have to set
CompatOS and CompatOsVersion manually in the config file.

The actual release can be found by parsing
/etc/upstream-release/lsb_release, see
https://unix.stackexchange.com/questions/270199