gshhs: Include all files in git repo (#1712)
This is the brute force tactics w r t the problems described in
bug #1712, basically that options like -DOCPN_BUNDLE_GSHHS=FULL
gives a late build error.

Other ways to fix this includes:

  - Automatically downloading missing data during build (not
    allowed in package builds).
  - Detecting missing data at an early stage, with feedback on
    how to download it.

This PR is filed with the idea is that the simplest solution is
to just include the data, adding some 73MB to the sources. Of
the 73MB 54MB are is the "full" files.