ci: flatpak: Adjust to only build flatpak on flatpak branch.
The idea is to push to the flatpak branch in a nightly job. The
flatpak branch takes a long time since it needs to rebuild
wxWidgets, and should generally be avoided in regular pushes
to master.

Using fine-grained conditions per branch  in travis.yml opens
up for different setups with for example either fast or complete
builds. This way, the travis setup has similar semantics as circleci.

Dropping the plug-mgr branch from set of built branches; it's a
leftover from before the plugin installer was merged.