build: Add a dist target (#1849)
Creates a dist target which builds a source tarball. Since this runs
after git commit it is able to add git version data (hash, tag and date)
to the tarball in the file git-version.

For the same reason it is also possible to use CI_BUILD on the
tarballs generated by 'make dist'.

Attempts to build from dirty git status are aborted. The target
requires that everything is committed.

If tar is available the tarball is repacked in order to avoid the
problems described in #1849. git and bash and presumed to exist.