build: Using OCPN_XXX MATCHES "ON" considered harmful (#1901).
Cmake variables are evaluated as true if 1, "TRUE", "YES", "ON". Users
expects that this is respected. Using conditions matching the variable
to "ON" limits the truth value to the single string "ON" which is
confusing and also inconsistent -- other variables are evaulated using
standard rules.

Apply standard evaluation to all variables. This also cleans up the code.