flatpak: flathub: Add a top-level manifest
Publishing the flatpak package on flathub.org requires the flatpak
manifest livse in the top-level dir. This manifest is almost
identical to the one in flatpak/ which is used to build snapshots.
However, it might diverge during the review process so having an
independent copy for now makes sense. When the dust settles the two
copies could probably be merged using some magic in flatpak/Makefile.