SendToGpsDlg: Handle missing Serial:/Usb: device prefix (#1957).
The combobox used to select commumication port allows uer to manually
enter the device name. This should include a Serial: ocr Usb: prefix,
like Serial:/dev/ttyUSB0.

This is far from obvious, and users will occasionally enter the plain
device name without prefix. As a temporary fix, add a Serial: prefix
in this situation.

A proper patch would remove the need for a prefix in the dialog and
add a Serial/Usb radio buttons control instead.