Added config_var wrapper.
Adding a light-weight listen/notify framework for global variables.

The problem addressed is to make it simple for GUI code to update when
a global variable is changed.

wxWidgets provides event messaging, which is a flexible, untyped way
to call any code. However, it has some drawbacks:

  - The event used must be declared ni headers, basically a global thing.

  - When sending event, the changing code must know the receiving window.

Especially the last bullet makes things complicated, requiring the
recieving window to be known and accessible in parameters, methods or
(horrors) yet another global when triggering a change.

The new framework for global variables on the sending/notifying side:

     ocpn::GlobalVar<wxString> compat_os(&g_compatOS);
     compat_os->set("ubuntu-gtk3-x86_64");

The listening part looks like:

     ocpn::GlobalVar<wxString> compat_os(&g_compatOS);

     // config sends a wxCommandEVent of type EVT_FOO to this on changes:
     wxDEFINE_EVENT(EVT_FOO, wxCommandEvent);
     compat_os->listen(this, EVT_FOO)

     // Handle  EVT_FOO as any event when it arrives, for example:
     Bind(EVT_FOO, [](wxCommandEvent&) { cout << "value has changed"; });

The config var variant works in a similar way besides the constructor.

     ocpn::ConfigVar<bool> expert("/PlugIns", "CatalogExpert");

Here, the get/set accessors provides more value, handling basically whatever
type is thrown upon them.