Allow more flexible regex filtering of NMEA messages
Background: on my vessel, there are more than one sources
emitting position information, for example $GPRMC and $IIRMC.
The latter provides rather poor accuracy and needs to be rejected,
otherwise the boat will "jump around" on the map.
But other data from this source like $IIDBT is of course needed.
With the currently implemented filtering I found it quite cumbersome
to find out the "bad guys on the bus" and do the job.

With my proposed extension, the filtering will be possible by
inserting in the section [Settings/NMEADataSource] simple
wildcard patterns like
;.II[GZ]..,.IIRM.;
at the location where the filter strings are stored.

The matching will only apply, if the pattern does not
have 2,3 or 5 characters.
This can be accomplished by inserting dots as wildcard characters.

of course, if my proposal is accepted, then the GUI settings dialogue
may be adapted to support this and prevent wrong input.

I have tested this modification on a Ubuntu Linux 20.04 machine feeding
recorded messages with gpsfake from the gpsd package.