isCompatible: Refactor and assume compile time setup (#2185).
Fix the issues mentioned in #2185. This includes removing the runtime
target detection. Instead, a new cmake option OCPN_TARGET_TUPLE
is added which hardwires the target platform.

OCPN_TARGET_TUPLE must be defined on all platforms where the
platform detection in configure time fails. This includes all
cross compiled platforms.

The syntax for OCPN_TARGET_TUPLE is "target;target-version;arch" for
example "flatpak-x86_64;18.08;x86_64". These corresponds to the elements
<target>, <target-version> and <target-arch> in the metadata file.
Using values not accepted in the XML metadata is not recommended.