Presimplify vector features before DuglasPeucker
Some IENC charts (Slovakia) contain extremely complex vector features with thousands of points
This leads to a very deep recursion in DouglasPeucker algorithm and a crash.

Add a preprocessing step using a naive pre-simplification procedure skipping proportional part of the points
regardless of their distance assuming that in an extremely complex shape with thousands of edges the probablity
of removing a significant feature is very low.