Improve the behavior of navobject property dialog on macOS (Closes #2317)
On macOS, when the modal dialog is shown on top of a non-modal window floating on top of the main application window, it may lose focus when the user switches to another application and back and there is no way to recover but knowing Esc closes the lost dialog or killing the application.
We sacrifice the non-modal window floating over the main app even when it does not have focus and switch to the more macOS native behavior of the window submerging and being recallable from the Window menu, using the window switching shortcuts (Cmd `) or Mission Control.