Added multitouch zoom and rotation
This adds multitouch zoom and rotation support (with 2 fingers)
This requires a wxWidgets version that has the support
for gestures, that is to say, a version greater than 3.1.1

The pan gesture is not usable (tested with wxWidgets 3.1.5),
there are a lot of missed events. So dragging is done the old way with
the motion events.

Additionnaly to zoom, the left long press makes the context
menu appear, and the doubletap resets the rotation to zero.

Tested on a Lenovo T480 with touchscreen lcd.

Signed-off-by: Thierry Bultel <tbultel@free.fr>