wxGesture: removed handling of 'pan' events
Pan events come additionnally to the MouseMove
events, that already perform the scrolling.
So when using them, the map slides twice as fast,
as bad side effect.
Moreover, pan events are stuck to 'X only' or
'Y only' moves, leading to bad user experience
when browsing the map.

This fix removes the handling of pan events

Signed-off-by: Thierry Bultel <tbultel@free.fr>