Better fix for #831870 using zip download.
The core issue in #831870 is a tar bug. Walk around the mess by
making a zip download instead which works out of the box.

I'm so sorry, this was sent to me in a patch which I cannot find,
and thus I cannot give credits to the submitter. I just borrowed
the basic idea which I remembered and re-implemented it.