ser_ports: Drop enumeration based on scanning /dev (#2422).
This code path is just not needed. libudev is ubiquitous on linux,
being part of the kernel interface. And for cases where libudev is
not available, mostly containers like flatpak, there is stll both
the sysfs-based alternative and the one using the library at
https://github.com/wjwwood/serial/blob/main/include/serial/serial.h.

Drop the code path completely, trust the three existing way for linux.
This means a new libudev-dev build dependency on Debian.