cmake: fix cpu feature detection for clang
While gcc will error when given unknown command line flags like -msse,
clang only warns and exits 0. This is problematic for try_compile()
because it expects the compile to fail when the flag passed is not
supported by the compiler.

I ran into this issue compiling opencpn for powerpc using clang where
it erroneosly detected sse.

The fix I came up with was to add -Werror to the try_compile command
lines which does not change the gcc behavior but does make clang
exit non-zero when passed a flag it does not support.