MouseZoom: Refactor, Scale UI values to user visible 1..100.
The actual , usable values for the zooming is 1.05 to 3.0. The upper
limit is somewhat arbitrary, the lower implied by the fact that 1.0
effectively disables mouse zoom.

However, from a user perspective these values makes no sense, a simple
1 - 100 scale seems more appropriate.

While on it, refactor the slider to a separate class to avoid adding too
many lines to a function already 300+ lines long.