Fix an address sanitizer error.
When the last bit to be unpacked happened to fall right on a byte
boundary, we could end up reading an extra byte and then masking
its value away. While normally fine, if that extra byte happened
to be off the end of the input buffer, this extra byte read would
read off the end of the array, potentially causing a segfault and
definitely causing an address sanitizer error.