Change pTrackList from being a wxList to a std::vector.
Only one place in the codebase meaningfully cares about
the difference, where we need to be careful about removing
elements from the vector.

Additionally, there's no reason to heap allocate the track
list when we're just going to store it in a global anyways.
Make the global by-value to remove one additional indirection.