ObservedVar: Add a listener class and events, clean up
Make clients use get_listener() rather then listen() so that
listening entities are removed on destruction.

Add a new events interface which can be used to carry data besides
just an empty notification.

Make more interfaces private, to make the intended usage clear.
Replace the wxWindow dependency with wxEvent, to shrink the interface
and make it possible to use also in CLI applications