Make g_BasePlatform a separate pointer (MS linker problems)
The original idea is a global variable OCPNPlatform* g_Platform.
Since the BasePlatfoem is a subclass of OCPNPlatform classes which
doesn not include OCPNPlatform can declare this as BasePlatform*
g_Platform.

This works fine on all platforms besides msvc, where the linker
refuses to link a BasePlatform* to a OCPNPlatform* object despite
that they are compatible.

Which means that a separate, global  g_BasePlatform must be used.
This is a pain, forcing us to keep g_BasePlatform and g_Platform in
sync.

The correct solution is to only have the g_BasePlatform global, using
a dynamic_cast<> when OCPNPlatform is required. However, this creates
a big delta which is hard to keep rebased for now.