Add methods to get an address from drivers.
In the case a plugin (or core) needs to send a message which not is
a reply it must deduce the destination address. This is done by
enumerating the drivers and select one of.

Add methods to generate an address to a given interface, possibly
also handling bus node addressing like n2k_name_t.