Use malloc/free to handle deleted pointer to GUI
plugin_loader will when a plugin is uninstalled first notify the GUI
and then delete the PlugInContainer. Since the notification is executed
onj next event cycle round it is executed after the container is
deleted.

Use a copy handled using  malloc/free to give the GUI a copy of the
deleted pointer so it can be removed from toolbars etc.

The correct solution is to use a shared_ptr instead.  However, this
will be a rather big patch which is better introduced after merging and
using the observable framework. Hence this dirty patch, which needs a
bug to be filed once merged.