observable: Clean up error detection
The basic assumption is that having two listeners with the same
(window, messagetype) is an error.

However, all class level listeners are initiated using a temporary
listener which is assigned to the class listener. During this window
there are two listeners with the same (window, messagetype). The
situation is rectified when the automatic variable goes out of scope.

For now, make the checks for duplicate listeners aware of this
situation and only log actual duplicates after unlisten() i. e.,
after the temporary variable destructor. This should catch
possible "real" duplicates. Also log when there are more than two
instances.

These checks are somewhat expensive, only run in debug mode.

This mess is a consequence of using value semantics for the listener
rather than the original unique_ptr. We should consider reverting this
change.