comm_drv_n2k_socketcan: Hide default constructor.
The comm_drv_n2k_socketcan uses std::enable_shared_from_this. This
means that all instances if this driver must be managed by a
shared_ptr, if not the behaviour is undefined.

Sooner or later this will bite us, so hide the default ctor and export
a static Create() factory function instead which returns a shared_ptr.