observable: Make sure each listener is unique.
A listener listening to a (window, event) tuple should not be possible
to copy since if there are two listeners to the same tuple we don't know
where to send the event when it occurs.

Read up on rval, lval and rvalue references like T&& and make sure that
the only possible way to transfer a listener is using std::move in which
case only the target listener will be active.