routeman: Use actual NavObjectChanges instance.
The assumption that there is only one NavObjectChanges instance, as
returned by NavObjectChanges::getInstance() is not completely true,
during initial setup there is a temporary instance.

Add a NavObjectChanges parameter to methods updating the database
so the correct instance can be used.

The basic problem here is that NavObjectChanges calls methods on
Routeman. The latter is an object above the database, so this call
should really be replaced by a signal or so to untangle the
spaghetti.