socketcan: Handle threading issues.
Make the worker thread code to live in a  class level variable.
Dynamically allocating this leads to multiple corner cases where
the thread might live on despite the variable is deleted. Besides,
the thread a raw pointer which basically is a bug per se.

Be careful when copying wxString from main thread to worker thread,
the default copying is not thread safe.

Use an atomic int as "run_flag", it is used by both main and worker
thread.

Fix up naming: a class is not a thread. Since the thread class now is
local in the implementation file there is no need for long globally
unique names.