AbstractCommDriver: Sort out SetListener() mess
The original signature of SetListener used a shared_ptr<DriverListener>
argument. This is wrong, each driver only has one listener, and makes
implementing SetListener() more or less impossible.

Use a sane DriverListener& argument instead. The only driver actually
implementing this is socketcan, but it should be easy when other
drivers are reviewed.