plugin_loader/manager: Fix horrible hack (#2744)
When deactivating a plugin, plugin_loader has emitted an event handled
by pluginmanager. However, since event handling is asynchronous this has
implies problem when the plugin to handle in pluginmanager has already
been removed by plugin_loader.

The hack has been a read-only, malloc'ed copy of the plugin, a
construction the compiler rightfully warns about since the plugin
container does not have a default ctor. This hack is horrible and
fragile.

Instead, retire the event and use a callback functin instead. This is
executed synchronously before the plugin is actually deleted and thus
solves the problem