plugin_loader, plugin_handler et al. Restore GetInstalled() semantics
Earlier patches to plugin_loader has broken PluginHandler::GetInstalled()
semantics. According to docs and history it should return a list of all
installed plugins but has been modified to return all entries in the GUI
list of plugins i. e. it returns also the catalog entries which
otherwise are in PluginHandler:.GetAvailable(). This breaks things when
trying to handle #3103.

Restore GetInstalled() to only return the installed plugins and let the
GUI parts create the complete list by adding GetInstalled() and
GetAvailable(). This removes some complicated parts, replacing it with
smaller functions.