opencpn-cmd: windows FTBFS: Handle msvc brain-dead linker
Contrary to other linkers out there the msvc toolchain linker cannot
handle polymorphism i. e. it cannot link a superclass pointer to a
derived one.

Sigh. Not the forst time stumbling into this, and probably not the last.
Why has windows a linker with a distinct smell of the 90'ies ?

On a sidenote the github CI job does not fail although it should.