build: ensure that API functions are available (#3578).
Part of the problems reported in #3528 is that the N2k API functions
available in the plugin API are not available in main opencpn. The
reason for that is that the linker has removed the comm_plugin_api.cpp
object file since it is not referenced.

The traditional way to handle this is to use the --whole-archive linker
option. However, there is no really clean way to apply this is cmake.

The alternative is to use a cmake OBJECT  library for the model. cmake
handles this as a collection of object files, and the linker gets not
chance to optimize awye anything.

However, this runs into another quirk: OBJECT libraries does not
implement transitive dependencies. For this reason all code using the
model need to link both to ocpn::model for the dependencies and to
ocpn::model-src for the OBJECT object files. This is not nice, but
perhaps the best possible until cmake 3.24 which has better handling
of --whole-archive