peer transfer: Shortcut version info
Using the new /api/get-version endpoint is problematic since it
is not implemented in 5.8. Hence teh only way to get the 5.8
version is to let the request for /api/get-version time out
and deem it as "old" in this case. Waiting fo rtimeout obviously
takes time.

Instead, use the /api/ping which originally just is for checking
the api key to also provide also the version. This means that a reply
without version info is considered 5.8. This scheme avoids the
timeout otherwise necessary.