rest_server: Avoid wxWin data in worker thread (#3588)
rest_server.cpp invokes methods affecting wxWindows data structures from
the server thread. Fix by moving to main thread.

This is a long standing issue which seems to be present also in 5.8.4
but for some reason does not trigger anything there.

Kudos to @hakansv which found the nasty #3588 bug while working
with #3584. And of course also to @bdbcat who found the basic problem.

Closes: #3588