Flatpak: Let some system D-Bus events pass through
We listen for org.freedesktop.systemd1 D-bus events in order to handle
USB devices which are inserted or ejected in the connection edit dialog.
Likewise, we listens to org.freedesktop.login1 in order to handle
suspend and resume signals

Make sure these signals are actually accessible in the sandbox using the
manifest finish-args configuration.