Fix assert failure due to using uninitialized wxMemoryDC
Don't call SetTextForeground() after calling SelectObject(wxNullBitmap)
on a wxMemoryDC, as the latter makes the DC unusable -- and changing the
text colour here seems completely unnecessary anyhow, as the DC is not
used any longer.

This prevents each frame redraw from resulting in an assert, which made
the plugin unusable when the asserts were active.