Avoid CMake error when configuring with OCPN_BUILD_SAMPLE=ON
Using target_link_libraries() with keyword arguments in the sample
plugin CMakefile was inconsistent with its use without them elsewhere.

Although it would be better to switch to the version with the keyword,
this would require many more changes, so for now just fix the plugin
build by removing the keyword from here too.